requirements.txt
<<<<<<< HEAD
git+https://github.com/Rapptz/discord.py
Pillow==8.4.0
psutil==5.8.0
requests==2.25.1
sentry_sdk==1.3.1
=======
git+https://github.com/Rapptz/discord.py
Pillow==8.4.0
psutil==5.8.0
requests==2.25.1
sentry_sdk==1.3.1
>>>>>>> 049ddcde2a090ba7492f82b75ee62cc010bbc290
toolbox_py==0.0.1